stubbed out move action client
